Refactor to remove go-colorful.
