+++ gradient/color.go
@@ -6,6 +6,7 @@ import (
 	"encoding/hex"
 	"errors"
 	"image/color"
+	"math"
 	"strings"
 )
 
@@ -20,5 +21,19 @@ func colorFromHex(h string) (color.RGBA, error) {
 		R: data[0],
 		G: data[1],
 		B: data[2],
+		A: math.MaxUint8,
 	}, nil
+}
+
+func linearInterpolate(first, second color.RGBA, pos float64) color.RGBA {
+	return color.RGBA{
+		R: lerp(first.R, second.R, pos),
+		G: lerp(first.G, second.G, pos),
+		B: lerp(first.B, second.B, pos),
+		A: lerp(first.A, second.A, pos),
+	}
+}
+
+func lerp(first, second uint8, stop float64) uint8 {
+	return uint8(float64(first)*(1.0-stop) + float64(second)*stop)
 }
