+++ gradient/gradient.go
@@ -4,17 +4,13 @@ package gradient
 
 import (
 	"image/color"
-	"math"
 
 	"github.com/ebeeton/buddhabrot-go/timer"
 	"github.com/go-playground/validator/v10"
-	"github.com/lucasb-eyer/go-colorful"
 )
 
-// This code used the gradient example from Lucas Beyer's excellent
-// go-colorful library.
 type stop struct {
-	col colorful.Color
+	col color.RGBA
 	pos float64
 }
 
@@ -27,7 +23,7 @@ func GetGradient(stops []Stop, count int) []color.RGBA {
 	t := gradientTable{}
 	p := []color.RGBA{}
 	for _, s := range stops {
-		c, err := colorful.Hex(s.Color)
+		c, err := colorFromHex(s.Color)
 		if err != nil {
 			panic("GetGradient: " + err.Error())
 		}
@@ -39,28 +35,19 @@ func GetGradient(stops []Stop, count int) []color.RGBA {
 	step := 1.0 / float64(count-1)
 	for i := 0; i < count; i++ {
 		col := t.getInterpolatedColor(float64(i) * step)
-		p = append(p, colorfulToColor(col))
+		p = append(p, col)
 	}
 	return p
 }
 
-func colorfulToColor(c colorful.Color) color.RGBA {
-	return color.RGBA{
-		R: uint8(math.MaxUint8 * c.R),
-		G: uint8(math.MaxUint8 * c.G),
-		B: uint8(math.MaxUint8 * c.B),
-		A: math.MaxUint8,
-	}
-}
-
-func (gt gradientTable) getInterpolatedColor(t float64) colorful.Color {
+func (gt gradientTable) getInterpolatedColor(t float64) color.RGBA {
 	for i := 0; i < len(gt)-1; i++ {
 		c1 := gt[i]
 		c2 := gt[i+1]
 		if c1.pos <= t && t <= c2.pos {
 			// Blend the two points we're between.
 			t := (t - c1.pos) / (c2.pos - c1.pos)
-			return c1.col.BlendRgb(c2.col, t).Clamped()
+			return linearInterpolate(c1.col, c2.col, t)
 		}
 	}
